# backend/api/api.yml
openapi: "3.0.0"
info:
  version: "1.20"
  title: Multicloud API
#produces:
#  - "application/json"
#basePath: "/api"
servers:
  - url: /api

# Paths supported by the server application
paths:
  /aws/vm:
    post:
      operationId: "aws.vm.createVM"
      tags:
        - "AWS-vm"
      summary: "Create an EC2 (VM)"
      description: "Create an EC2 (VM)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        200:
          description: "Successful created VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "aws.vm.loadAllVM"
      tags:
        - "AWS-vm"
      summary: "Loads All VMs"
      description: "Loads all VMS"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful loaded all VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponseFullist'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "aws.vm.deleteVM"
      tags:
        - "AWS-vm"
      summary: "Deletes an EC2 (VM)"
      description: "Deletes an EC2 (VM)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequestShort'
      responses:
        200:
          description: "Successful deleted VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /aws/vm/{instanceId}:
    post:
      operationId: "aws.vm.getVM"
      tags:
        - "AWS-vm"
      summary: "Get the VM based on provided instanceId"
      description : "Get the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful read VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "aws.vm.stopVM"
      tags:
        - "AWS-vm"
      summary: "stop the VM based on provided instanceId"
      description: "stop the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful stopped VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "aws.vm.startVM"
      tags:
        - "AWS-vm"
      summary: "Starts the VM based on provided instanceId"
      description: "Starts the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful started VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
  /aws/keypair:
    post:
      operationId: "aws.keypair.loadOrCreateKeyPair"
      tags:
        - "AWS-keypair"
      summary: "Create or Load a Keypair"
      description: "Creates or loads the provided keypair"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyPairRequest'
      responses:
        200:
          description: "Successful read or created keypair"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPairResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "aws.keypair.getAllKeypairs"
      tags:
        - "AWS-keypair"
      summary: "Loads all Keypairs"
      description: "Loads all Keypairs"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful loaded all keypairs"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPairResponseFullList'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "aws.keypair.deleteKeyPair"
      tags:
        - "AWS-keypair"
      summary: "Deletes a Keypair"
      description: "Deletes the provided keypair"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyPairRequest'
      responses:
        200:
          description: "Successful deleted keypair"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
  /azure/network:
    post:
      operationId: "azure.network.createANetwork"
      tags:
        - "AZURE-network"
      summary: "Creates a Network"
      description: "Creates a Network"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkRequest'
      responses:
        200:
          description: "Successful created network"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "azure.network.getAllNetworks"
      tags:
        - "AZURE-network"
      summary: "gets all Networks"
      description: "gets all Networks"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkListRequest'
      responses:
        200:
          description: "Successful created network"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "azure.network.deleteNetwork"
      tags:
        - "AZURE-network"
      summary: "Deletes a Network"
      description: "Deletes a Network"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteNetworkRequest'
      responses:
        200:
          description: "Successful deleted Network"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /azure/network/{networkName}:
    patch:
      operationId: "azure.network.getANetwork"
      tags:
        - "AZURE-network"
      summary: "Get a network from azure"
      description: "Get a network from azure"
      parameters:
        - $ref: '#/components/parameters/networkName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkListRequest'
      responses:
        200:
          description: "Successful read Network"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /azure/subnet:
    post:
      operationId: "azure.network.createOrUpdateSubnet"
      tags:
        - "AZURE-network"
      summary: "Creates or Updates a Subnet"
      description: "Creates or Updates a Subnet"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubnetRequest'
      responses:
        200:
          description: "Successful created subnet"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subnet'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /azure/nic:
    post:
      operationId: "azure.network.CreateOrUpdateNic"
      tags:
        - "AZURE-network"
      summary: "Creates or Updates a NIC"
      description: "Creates or Updates a NIC"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NICRequest'
      responses:
        200:
          description: "Successful created NIC"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NICResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "azure.network.getALLNIC"
      tags:
        - "AZURE-network"
      summary: "get all NICs"
      description: "get all NICs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkListRequest'
      responses:
        200:
          description: "Successful read NIC"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NICResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
  /azure/vm:
    post:
      operationId: "azure.vm.createaVM"
      tags:
        - "AZURE-vm"
      summary: "Creates a VM"
      description: "Creates a VM"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AzureVMRequest'
      responses:
        200:
          description: "Successful created VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AzureVMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "azure.vm.getAllVMS"
      tags:
        - "AZURE-vm"
      summary: "gets all VMs"
      description: "gets all VMs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkListRequest'
      responses:
        200:
          description: "Successful read VMs"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMListResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "azure.vm.deleteVM"
      tags:
        - "AZURE-vm"
      summary: "deletes a vm from azure"
      description: "deletes a vm from azure"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMSmallRequest'
      responses:
        200:
          description: "Successful deleted VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "azure.vm.startVM"
      tags:
        - "AZURE-vm"
      summary: "starts a VM"
      description: "starts a VM"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMSmallRequest'
      responses:
        200:
          description: "Successful started VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
components:
  parameters:
...
    networkName:
      name: networkName
      in: path
      required: true
      description: networkName
      schema:
        type: string
    instanceId:
      name: instanceId
      in: path
      required: true
      description: InstanceId
      schema:
        type: string
...
  schemas:
    TokenRequest:
      type: object
      properties:
        token:
          type: string
    NotAuthorizedResponse:
      type: object
      properties:
        detail:
          type: string
        status:
          type: number
        title:
          type: string
        type:
          type: string
...
    VMSmallRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vmname:
          type: string
    VMListResponse:
      type: array
      items:
        $ref: '#/components/schemas/AzureVMResponse'
...
    AzureVMRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vmname:
          type: string
        username:
          type: string
        password:
          type: string
        vmsize:
          type: string
        publishername:
          type: string
        offername:
          type: string
        skuname:
          type: string
        version:
          type: string
        nicid:
          type: string
...
    NICResponse:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        virtual_machine:
          type: object
          properties:
            id:
              type: string
        private_endpoint:
          type: object
          properties:
            name:
              type: string
            instanceType:
              type: string
            provisioning_state:
              type: string
        hosted_workloads:
          type: array
          items:
            type: string
    NICRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        nicname:
          type: string
        ipconfigname:
          type: string
        subnetid:
          type: string
        vnetName:
          type: string
    SubnetRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vnetName:
          type: string
        subnetName:
          type: string
        subnetIp:
          type: string
    DeleteNetworkRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vnetName:
          type: string
    NetworkListRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
    NetworkRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vnetName:
          type: string
        networkIp:
          type: string
    NetworkResponse:
      type: object
      properties:
        id:
          type: string
        location:
          type: string
        tags:
          type: object
          properties:
            key:
              type: string
            value:
              type: string
        address_space:
          type: object
          properties:
            address_prefixes:
              type: array
              items:
                type: string
        dhcp_options:
          type: object
          properties:
            dns_servers:
              type: array
              items:
                type: string
        subnets:
          type: array
          items:
            $ref: '#/components/schemas/Subnet'
        virtual_network_peerings:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              allow_virtual_network_access:
                type: boolean
              allow_forwarded_traffic:
                type: boolean
              allow_gateway_transit:
                type: boolean
              use_remote_gateways:
                type: boolean
              remote_virtual_network:
                type: object
                properties:
                  id:
                    type: string
              remote_address_space:
                type: object
                properties:
                  address_prefixes:
                    type: array
                    items:
                      type: string
              peering_state:
                type: string
              name:
                type: string
        enable_ddos_protection:
          type: boolean
        enable_vm_protection:
          type: boolean
        ddos_protection_plan:
          type: object
          properties:
            id:
              type: string
        bgp_communities:
          type: object
          properties:
            virtual_network_community:
              type: string
    Subnet:
      type: object
      properties:
        id:
          type: string
        address_prefix:
          type: string
        address_prefixes:
          type: array
          items:
            type: string
        network_security_group:
          type: object
          properties:
            id:
              type: string
            location:
              type: string
            tags:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
            security_rules:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  description:
                    type: string
                  protocol:
                    type: string
                  source_port_range:
                    type: string
                  destination_port_range:
                    type: string
                  source_address_prefix:
                    type: string
                  source_address_prefixes:
                    type: array
                    items:
                      type: string
                  source_application_security_groups:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        location:
                          type: string
                        tags:
                          type: object
                          properties:
                            key:
                              type: string
                            value:
                              type: string
                  destination_address_prefix:
                    type: string
                  destination_address_prefixes:
                    type: array
                    items:
                      type: string
                  destination_application_security_groups:
                    type: array
                    items:
                      type: string
                  source_port_ranges:
                    type: array
                    items:
                      type: string
                  destination_port_ranges:
                    type: array
                    items:
                      type: string
                  access:
                    type: string
                  priority:
                    type: number
                  direction:
                    type: string
                  name:
                    type: string
        route_table:
          type: object
          properties:
            name:
              type: string
            type:
              type: string
            provisioning_state:
              type: string
            etag:
              type: string
        nat_gateway:
          type: object
          properties:
            id:
              type: string
        service_endpoints:
          type: array
          items:
            type: object
            properties:
              provisioning_state:
                type: string
        service_endpoint_policies:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              resource_guid:
                type: string
              provisioning_state:
                type: string
              etag:
                type: string
        delegations:
          type: array
          items:
            type: object
            properties:
              actions:
                type: array
                items:
                  type: string
              provisioning_state:
                type: string
              etag:
                type: string
        private_endpoint_network_policies:
          type: string
        private_link_service_network_policies:
          type: string
        name:
          type: string
...
    VMRequest:
      type: object
      properties:
        token:
          type: string
        imageId:
          type: string
        instanceType:
          type: string
        keyName:
          type: string
        securityGroups:
          type: array
          items:
            type: string
        vmname:
          type: string
    VMRequestShort:
      type: object
      properties:
        token:
          type: string
        instanceId:
          type: string
    VMResponse:
      type: object
      properties:
        instanceId:
          type: string
    State:
      type: object
      properties:
        State:
          type: string
...
    VMResponseFullist:
      type: object
      properties:
        instances:
          type: array
          items:
            $ref: '#/components/schemas/VMResponseFull'
...
    KeyPairResponseFullList:
      type: object
      properties:
        keypairs:
          type: array
          items:
            $ref: '#/components/schemas/KeyPairResponseFull'
    KeyPairResponseFull:
      type: object
      properties:
        KeyPairId:
          type: string
        KeyFingerprint:
          type: string
        KeyName:
          type: string
        Tags:
          type: array
          items:
            type: object
            properties:
              Key:
                type: string
              Value:
                type: string
    KeyPairRequest:
      type: object
      properties:
        token:
          type: string
        keypair:
          type: string
    KeyPairResponse:
      type: object
      properties:
        keypairId:
          type: string
...